requirements.txt
python-telegram-bot==13.7
gspread==5.7.2
oauth2client==4.1.3
apscheduler==3.6.3
pytz==2022.7.1
python-dotenv==1.0.0
